requirements.txt

aiogram>=2.20
requests>=2.28.1
Pillow>=9.3.0
rembg>=2.0.60
